+++ app/_components/AddressSection/StakeSection.tsx
@@ -1,7 +1,7 @@
 import type { Output } from "~/napi-pallas";
-import { PropBlock, Section, EmptyBlock } from "../DissectedTx/Constructors";
+import { PropBlock, Section } from "../Constructors";
 
-export function ShelleySection(props: { data: Output }) {
+export function StakeSection(props: { data: Output }) {
   const { data } = props;
 
   return (
@@ -14,12 +14,11 @@ export function ShelleySection(props: { data: Output }) {
           className="text-blue-700 underline hover:text-blue-500"
           href="https://cips.cardano.org/cip/CIP-0019"
           target="_blank"
-          rel="noopener noreferrer"
         >
           CIP-0019
         </a>
         . The CIP explains that there are 3 types of possible address, each one
-        following a different encoding format: Shelley, Stake, or Byron.
+        following a different encoding format: Shelley, Stake or Byron.
       </p>
       <PropBlock
         color="green"
@@ -29,42 +28,20 @@ export function ShelleySection(props: { data: Output }) {
       <Section title="Parsed Address">
         <p className="text-xl text-gray-600">
           The address entered is of type&nbsp;
-          <code>Shelley</code>. Shelley addresses contain three pieces of
-          information: network id, payment part, and a delegation part.
+          <code>Stake</code>. Stake addresses contain two pieces of information:
+          network tag and delegation info.
         </p>
         <PropBlock title="type" value={data?.address?.kind} />
-        <Section title="Network Id">
+        <Section title="Network Tag">
           <p className="text-xl text-gray-600">
-            The network id is a flag to indicate to which network it belongs
+            The netword tag is a flag to indicate to which network it belongs
             (either mainnet or a testnet).
           </p>
-          <PropBlock title="network id" value={data?.address?.network} />
+          <PropBlock title="network tag" value={data?.address?.network} />
         </Section>
-        {!!data.address?.paymentPart && (
-          <Section title="Payment Part">
-            <p className="text-xl text-gray-600">
-              The payment part describes who has control of the ownership of the
-              locked values. There are two options: a verification key or a
-              script. The address includes a flag to differentiate the two.
-            </p>
-            <PropBlock
-              title="kind"
-              value={
-                data.address.paymentPart.isScript
-                  ? "script"
-                  : "verification key"
-              }
-            />
-            <PropBlock
-              color="green"
-              title="hash"
-              value={data.address.paymentPart.hash}
-            />
-          </Section>
-        )}
         {(!!data.address?.delegationPart?.hash ||
           !!data.address?.delegationPart?.pointer) && (
-          <Section title="Delegation Part">
+          <Section title="Delegation Info">
             <p className="text-xl text-gray-600">
               The delegation part describes who has control of the staking of
               the locked values. There are two options: a verification key or a
@@ -94,18 +71,6 @@ export function ShelleySection(props: { data: Output }) {
             )}
           </Section>
         )}
-        {!data.address?.delegationPart?.hash &&
-          !data.address?.delegationPart?.pointer && (
-            <Section title="Delegation Part">
-              <p className="text-xl text-gray-600">
-                The delegation part describes who has control of the staking of
-                the locked values. This address doesn&apos;t specify a
-                delegation part, which means there&apos;s no way to delegate the
-                locked values of this address.
-              </p>
-              <EmptyBlock />
-            </Section>
-          )}
       </Section>
     </Section>
   );
