+++ app/_components/index.ts
@@ -1,13 +1,15 @@
+export { setCBOR } from "~/app/_components/Header/header.helper";
+export { ByronSection, ShelleySection, StakeSection } from "./AddressSection";
 export { Button } from "./Button";
-export * from "./DissectTx/Constructors";
-export { DissectSection } from "./DissectTx/DissectSection";
+export * from "./Constructors";
+export { DissectSection } from "./DissectSection/DissectSection";
+export { DslSection } from "./DslSection/DslSection";
 export { Error } from "./Error";
-export { Examples } from "./Examples";
+export { Examples, ExamplesAddress } from "./Examples";
 export { Header } from "./Header/Header";
 export { InfoPanel, TxInfo, UtxoInfo } from "./InfoPanel";
-export { CustomInput as Input } from "./Input";
+export { AddressInput } from "./Input/AddressInput";
+export { CustomInput as Input } from "./Input/Input";
+export { TxInput } from "./Input/TxInput";
 export { Playground } from "./Playground";
 export { Line, Transaction, Utxo } from "./Transaction";
-export { TxInput } from "./TxInput";
-export { AddressInput } from "./AddressInput";
-export { ByronSection, ShelleySection, StakeSection } from "./Address";
