+++ app/api/hash/route.ts
@@ -2,7 +2,7 @@ import { ReasonPhrases, StatusCodes } from "http-status-codes";
 import { type NextRequest } from "next/server";
 import { ZodError } from "zod";
 import { getQuery, getTxFromHashSchema } from "~/app/_utils";
-import { hashHandler } from "../_handlers";
+import { hashHandler } from "~/app/api/_handlers";
 
 export async function GET(req: NextRequest) {
   try {
